01-core-systems/git-internals-master/requirements.txt
gitpython>=3.1.40
polars>=0.20.0
pyarrow>=14.0.0
